requirements.txt
ultralytics           # YOLOv8 library
opencv-python         # image I/O
Pillow                # optional, for image sizes in conversion
pytest